Latest Version

+++ home.py
@@ -1,4 +1,4 @@
-from contextlib import nullcontext
+from logging import NullHandler
 from tkinter import *
 from tkinter import ttk
 from DB import *
@@ -13,8 +13,7 @@ def on_key_press(event):
     if event.keysym == "Escape":
         root.destroy()
         print("Page Closed")
-    else:
-        print(event.keysym)
+
 
 def update_time():
     label_date.config(text =f"{datetime.now().strftime("%Y-%m-%d")}")
@@ -23,20 +22,28 @@ def update_time():
 
 search_item = "name"  #--important variable--
 table_object_Selected = 0 #--important variable--
-
+search_input = ''  #--important variable--
 search_entry_placeholder = f"Enter {search_item}"
-def search_entry_on_focus_in(event):
-    if search_entry.get() == search_entry_placeholder:
-        search_entry.delete(0, END)  # Clear the placeholder when the entry gets focus
-def search_entry_on_mouse_enter(event):
-    if search_entry.get() == search_entry_placeholder:
-        search_entry.delete(0, END)  # Clear the placeholder when the entry gets focus
-def search_entry_on_mouse_leave(event):
-    if search_entry.get() == "":
-        search_entry.insert(0, search_entry_placeholder)  # Reinsert the placeholder text when focus is lost
+
 def search_entry_on_key_press(event):
-    if search_entry.get() == search_entry_placeholder:
-        search_entry.delete(0, END)  # Clear the placeholder when the entry gets focus
+    global search_input
+    if event.char.isalpha() or event.keysym == "space" or event.char.isnumeric():
+        search_input = search_entry.get() + event.char
+        # print(event.char)
+        print(search_input)
+        print(f"keysym :{event.keysym}")
+        search_result_update(search_input)
+    elif event.keysym == "Return":
+        search_result_update(search_input)
+
+
+
+def search_entry_on_backspace(event):
+    global search_input
+    search_input = search_entry.get()[:-1] #to remove last letter
+    print(search_input)
+    search_result_update(search_input)
+
 def on_item_click(event):
     table_region = table.identify_region(event.x, event.y)
     if table_region == "cell":
@@ -44,36 +51,47 @@ def on_item_click(event):
         # Get the item row number
         item_id = table.identify_row(event.y)
         table_object_Selected = table.item(item_id)['values'][1]
-#variable to choose what we will search for
+
 def search_item_buttons_handler(option):
     global search_item
-    global search_entry_placeholder
     search_item = option
     search_entry.delete(0, END)  # Clear the placeholder when the entry gets focus
-    search_entry_placeholder = f"Enter {search_item}"
-    search_entry.insert(0, search_entry_placeholder)
     entry_search_name.config(bg="lightblue" if option == "name" else "SystemButtonFace")
     entry_search_phone.config(bg="lightblue" if option == "phone number" else "SystemButtonFace")
     entry_search_intake.config(bg="lightblue" if option == "intake" else "SystemButtonFace")
     entry_search_id.config(bg="lightblue" if option == "id" else "SystemButtonFace")
 
+def search_result_update(input_search):
+    search_result = search_db(search_item,input_search)
+    for item in table.get_children():
+        table.delete(item)
+    numbered_data = [(i + 1, *row) for i, row in enumerate(search_result)]
+    for i in numbered_data:
+        table.insert("", END, values=i)
+
 def update_table():
     for item in table.get_children():
         table.delete(item)
-    data = get_data('SELECT ID, Name, Phone_Number, "Intake_No" FROM students')
+    data = get_data('SELECT ID, Name, Phone_Number, "Intake_No" FROM Students')
     numbered_data = [(i + 1, *row) for i, row in enumerate(data)]
     for i in numbered_data:
         table.insert("", END, values=i)
 
 def delete_user():
-    #global table_object_Selected
+    global table_object_Selected
+    if table_object_Selected == 0:
+        messagebox.showerror("Select user","Select a user first !")
+        return 0
     user_data = get_by_id(table_object_Selected)
     response = messagebox.askyesno("Confirmation", f"Are you sure you want to delete user :{user_data[1]}?")
     if response:
         delete_by_id(table_object_Selected)
         messagebox.showinfo('Done',"user deleted successfully !!")
+        update_table()
+        table_object_Selected = 0
     else:
         pass
+
 
 root = Tk()
 root.title("Home")
@@ -97,22 +115,18 @@ update_time()
 ######################  Search Button & handling its events    ######################
 search_entry = Entry(root,width = 50,font = "arial 15")
 search_entry.place(relx=0.5, y = 80 ,anchor='center')
-# Bind events to handle the focus in and out behavior
-search_entry.bind("<FocusIn>", search_entry_on_focus_in)
+# Bind event to handle the key press
 search_entry.bind("<Key>", search_entry_on_key_press)
-search_entry.bind("<Enter>", search_entry_on_mouse_enter)
-search_entry.bind("<Leave>", search_entry_on_mouse_leave)
-
-search_entry.insert(0, search_entry_placeholder)
+search_entry.bind("<BackSpace>", search_entry_on_backspace)
 ######################################################################################
 
 
 ######################  Search Buttons    ######################
 Label(root,text = "Search By:",bg ="#06275c",font ="arial 8",fg ="white").place(x = 491,y=100)
 entry_search_name = Button(root, text="Name",font = "Arial 8",height=1,command=lambda:search_item_buttons_handler("name"))
-entry_search_phone = Button(root, text="Phone Number",command=lambda:search_item_buttons_handler("phone number"))
-entry_search_intake = Button(root, text="Intake", command=lambda:search_item_buttons_handler("intake"))
-entry_search_id = Button(root, text="ID", command=lambda:search_item_buttons_handler("id"))
+entry_search_phone = Button(root, text="Phone Number",command=lambda:search_item_buttons_handler("Phone_Number"))
+entry_search_intake = Button(root, text="Intake", command=lambda:search_item_buttons_handler("Intake_No"))
+entry_search_id = Button(root, text="ID", command=lambda:search_item_buttons_handler("ID"))
 
 entry_search_name.place(x = 561,y = 100)
 entry_search_phone.place(x = 607,y = 100)
@@ -151,7 +165,7 @@ table.bind("<ButtonRelease-1>", on_item_click) #to return the user number select
 ######################  CRUD operations buttons    ######################
 button_create = Button(text = "Create",font="arial 30",fg="#06275c",bg="#eed055",command =lambda: display_create(root))
 button_Details = Button(text = "Details",font="arial 30",fg="#06275c",bg="#eed055",command=lambda: display_details(table_object_Selected,root))
-button_Update = Button(text = "Update",font="arial 30",fg="#06275c",bg="#eed055",command= lambda:display_update(root,table_object_Selected))
+button_Update = Button(text = "Update",font="arial 30",fg="#06275c",bg="#eed055",command=lambda:display_update(root, table_object_Selected))
 button_Delete = Button(text = "Delete",font="arial 30",fg="#06275c",bg="#eed055",command = delete_user)
 button_create.place(x=450,y=680,anchor = "center")
 button_Details.place(x=662,y=680,anchor = "center")
@@ -159,20 +173,19 @@ button_Update.place(x=874,y=680,anchor = "center")
 button_Delete.place(x=1086,y=680,anchor = "center")
 #########################################################################
 
-
-
 ######################  Reload Table Button    ######################
 reload_icon = ImageTk.PhotoImage(Image.open("Images/reload.png").resize((50,50)))
 button_reload = Button(text = "Reload",font="arial 30",fg="#06275c",bg="#eed055",image = str(reload_icon),command =update_table)
 button_reload.place(x=1180,y=210)
 #######################################################################
 
+button_about_me = Button(text = "About",font="arial 30",fg="#06275c",bg="#eed055")
 
 ######################  Load DataBase Data    ######################
 update_table()
 ####################################################################
-
-
-
+for i in range(210,310):
+    root.after(10,lambda i:i+1)
+    button_reload.place(x=1180,y=i)
 root.mainloop()
 root.resizable(False, False)
